Update project structure and add Cognitive Load Plot
- Refactor project directory structure to align with Next.js 13+ app directory conventions:
  - Moved components to /components directory
  - Moved custom hooks to /hooks directory
  - Moved utility functions to /utils directory
  - Ensured all imports are updated to reflect new structure

- Implement dynamic imports for Plotly to avoid SSR issues:
  - Updated ActionsPlot.tsx and CognitiveLoadPlot.tsx to use dynamic imports with { ssr: false }

- Added Cognitive Load Plot component:
  - Created CognitiveLoadPlot component in /components/plots
  - Implemented useCognitiveLoadData custom hook to fetch and process cognitive load data
